refactor: Optimize Blocks marshaling with embedding
